Fix wrong parameterization

+++ src/main/java/ar/com/kfgodel/function/intertype/bytes/ByteToBooleanFunction.java
@@ -5,13 +5,13 @@ import java.util.function.Function;
 /**
  * Created by kfgodel on 23/07/17.
  */
-public interface ByteToBooleanFunction extends Function<Integer, Boolean> {
+public interface ByteToBooleanFunction extends Function<Byte, Boolean> {
 
   boolean apply(byte aBoolean);
 
   @Override
-  default Boolean apply(Integer aBoolean){
-    return this.apply(aBoolean.intValue());
+  default Boolean apply(Byte aBoolean){
+    return this.apply(aBoolean.byteValue());
   }
 
 }
