+++ src/main/java/ar/com/kfgodel/function/intertype/bytes/ByteToByteFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface ByteToByteFunction extends Function<Integer, Byte> {
+public interface ByteToByteFunction extends Function<Byte, Byte> {
 
   byte apply(byte input);
 
   @Override
-  default Byte apply(Integer input){
-    return apply(input.intValue());
+  default Byte apply(Byte input){
+    return apply(input.byteValue());
   }
 }
