+++ src/main/java/ar/com/kfgodel/function/intertype/bytes/ByteToCharFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface ByteToCharFunction extends Function<Integer, Character> {
+public interface ByteToCharFunction extends Function<Byte, Character> {
 
   char apply(byte input);
 
   @Override
-  default Character apply(Integer input){
-    return apply(input.intValue());
+  default Character apply(Byte input){
+    return apply(input.byteValue());
   }
 }
