+++ src/main/java/ar/com/kfgodel/function/intertype/bytes/ByteToFloatFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface ByteToFloatFunction extends Function<Integer, Float> {
+public interface ByteToFloatFunction extends Function<Byte, Float> {
 
   float apply(byte input);
 
   @Override
-  default Float apply(Integer input){
-    return apply(input.intValue());
+  default Float apply(Byte input){
+    return apply(input.byteValue());
   }
 }
