+++ src/main/java/ar/com/kfgodel/function/intertype/bytes/ByteToIntFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface ByteToIntFunction extends Function<Integer, Integer> {
+public interface ByteToIntFunction extends Function<Byte, Integer> {
 
   int apply(byte input);
 
   @Override
-  default Integer apply(Integer input){
-    return apply(input.intValue());
+  default Integer apply(Byte input){
+    return apply(input.byteValue());
   }
 }
