+++ src/main/java/ar/com/kfgodel/function/intertype/bytes/ByteToLongFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface ByteToLongFunction extends Function<Integer, Long> {
+public interface ByteToLongFunction extends Function<Byte, Long> {
 
   long apply(byte input);
 
   @Override
-  default Long apply(Integer input){
-    return apply(input.intValue());
+  default Long apply(Byte input){
+    return apply(input.byteValue());
   }
 }
