+++ src/main/java/ar/com/kfgodel/function/intertype/bytes/ByteToObjectFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface ByteToObjectFunction<T> extends Function<Integer, T> {
+public interface ByteToObjectFunction<T> extends Function<Byte, T> {
 
   T apply(byte input);
 
   @Override
-  default T apply(Integer input){
-    return apply(input.intValue());
+  default T apply(Byte input){
+    return apply(input.byteValue());
   }
 }
