+++ src/main/java/ar/com/kfgodel/function/intertype/bytes/ByteToShortFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface ByteToShortFunction extends Function<Integer, Short> {
+public interface ByteToShortFunction extends Function<Byte, Short> {
 
   short apply(byte input);
 
   @Override
-  default Short apply(Integer input){
-    return apply(input.intValue());
+  default Short apply(Byte input){
+    return apply(input.byteValue());
   }
 }
