+++ src/main/java/ar/com/kfgodel/function/intertype/doubles/DoubleToBooleanFunction.java
@@ -5,13 +5,13 @@ import java.util.function.Function;
 /**
  * Created by kfgodel on 23/07/17.
  */
-public interface DoubleToBooleanFunction extends Function<Integer, Boolean> {
+public interface DoubleToBooleanFunction extends Function<Double, Boolean> {
 
   boolean apply(double aBoolean);
 
   @Override
-  default Boolean apply(Integer aBoolean){
-    return this.apply(aBoolean.intValue());
+  default Boolean apply(Double aBoolean){
+    return this.apply(aBoolean.doubleValue());
   }
 
 }
