+++ src/main/java/ar/com/kfgodel/function/intertype/doubles/DoubleToDoubleFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface DoubleToDoubleFunction extends Function<Integer, Double> {
+public interface DoubleToDoubleFunction extends Function<Double, Double> {
 
   double apply(double input);
 
   @Override
-  default Double apply(Integer input){
-    return apply(input.intValue());
+  default Double apply(Double input){
+    return apply(input.doubleValue());
   }
 }
