+++ src/main/java/ar/com/kfgodel/function/intertype/doubles/DoubleToIntFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface DoubleToIntFunction extends Function<Integer, Integer> {
+public interface DoubleToIntFunction extends Function<Double, Integer> {
 
   int apply(double input);
 
   @Override
-  default Integer apply(Integer input){
-    return apply(input.intValue());
+  default Integer apply(Double input){
+    return apply(input.doubleValue());
   }
 }
