+++ src/main/java/ar/com/kfgodel/function/intertype/doubles/DoubleToLongFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface DoubleToLongFunction extends Function<Integer, Long> {
+public interface DoubleToLongFunction extends Function<Double, Long> {
 
   long apply(double input);
 
   @Override
-  default Long apply(Integer input){
-    return apply(input.intValue());
+  default Long apply(Double input){
+    return apply(input.doubleValue());
   }
 }
