+++ src/main/java/ar/com/kfgodel/function/intertype/floats/FloatToBooleanFunction.java
@@ -5,13 +5,13 @@ import java.util.function.Function;
 /**
  * Created by kfgodel on 23/07/17.
  */
-public interface FloatToBooleanFunction extends Function<Integer, Boolean> {
+public interface FloatToBooleanFunction extends Function<Float, Boolean> {
 
   boolean apply(float aBoolean);
 
   @Override
-  default Boolean apply(Integer aBoolean){
-    return this.apply(aBoolean.intValue());
+  default Boolean apply(Float aBoolean){
+    return this.apply(aBoolean.floatValue());
   }
 
 }
