+++ src/main/java/ar/com/kfgodel/function/intertype/floats/FloatToCharFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface FloatToCharFunction extends Function<Integer, Character> {
+public interface FloatToCharFunction extends Function<Float, Character> {
 
   char apply(float input);
 
   @Override
-  default Character apply(Integer input){
-    return apply(input.intValue());
+  default Character apply(Float input){
+    return apply(input.floatValue());
   }
 }
