+++ src/main/java/ar/com/kfgodel/function/intertype/floats/FloatToDoubleFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface FloatToDoubleFunction extends Function<Integer, Double> {
+public interface FloatToDoubleFunction extends Function<Float, Double> {
 
   double apply(float input);
 
   @Override
-  default Double apply(Integer input){
-    return apply(input.intValue());
+  default Double apply(Float input){
+    return apply(input.floatValue());
   }
 }
