+++ src/main/java/ar/com/kfgodel/function/intertype/floats/FloatToFloatFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface FloatToFloatFunction extends Function<Integer, Float> {
+public interface FloatToFloatFunction extends Function<Float, Float> {
 
   float apply(float input);
 
   @Override
-  default Float apply(Integer input){
-    return apply(input.intValue());
+  default Float apply(Float input){
+    return apply(input.floatValue());
   }
 }
