+++ src/main/java/ar/com/kfgodel/function/intertype/floats/FloatToIntFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface FloatToIntFunction extends Function<Integer, Integer> {
+public interface FloatToIntFunction extends Function<Float, Integer> {
 
   int apply(float input);
 
   @Override
-  default Integer apply(Integer input){
-    return apply(input.intValue());
+  default Integer apply(Float input){
+    return apply(input.floatValue());
   }
 }
