+++ src/main/java/ar/com/kfgodel/function/intertype/floats/FloatToLongFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface FloatToLongFunction extends Function<Integer, Long> {
+public interface FloatToLongFunction extends Function<Float, Long> {
 
   long apply(float input);
 
   @Override
-  default Long apply(Integer input){
-    return apply(input.intValue());
+  default Long apply(Float input){
+    return apply(input.floatValue());
   }
 }
