+++ src/main/java/ar/com/kfgodel/function/intertype/floats/FloatToObjectFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface FloatToObjectFunction<T> extends Function<Integer, T> {
+public interface FloatToObjectFunction<T> extends Function<Float, T> {
 
   T apply(float input);
 
   @Override
-  default T apply(Integer input){
-    return apply(input.intValue());
+  default T apply(Float input){
+    return apply(input.floatValue());
   }
 }
