+++ src/main/java/ar/com/kfgodel/function/intertype/floats/FloatToShortFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface FloatToShortFunction extends Function<Integer, Short> {
+public interface FloatToShortFunction extends Function<Float, Short> {
 
   short apply(float input);
 
   @Override
-  default Short apply(Integer input){
-    return apply(input.intValue());
+  default Short apply(Float input){
+    return apply(input.floatValue());
   }
 }
