+++ src/main/java/ar/com/kfgodel/function/intertype/longs/LongToBooleanFunction.java
@@ -5,13 +5,13 @@ import java.util.function.Function;
 /**
  * Created by kfgodel on 23/07/17.
  */
-public interface LongToBooleanFunction extends Function<Integer, Boolean> {
+public interface LongToBooleanFunction extends Function<Long, Boolean> {
 
   boolean apply(long aBoolean);
 
   @Override
-  default Boolean apply(Integer aBoolean){
-    return this.apply(aBoolean.intValue());
+  default Boolean apply(Long aBoolean){
+    return this.apply(aBoolean.longValue());
   }
 
 }
