+++ src/main/java/ar/com/kfgodel/function/intertype/longs/LongToDoubleFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface LongToDoubleFunction extends Function<Integer, Double> {
+public interface LongToDoubleFunction extends Function<Long, Double> {
 
   double apply(long input);
 
   @Override
-  default Double apply(Integer input){
-    return apply(input.intValue());
+  default Double apply(Long input){
+    return apply(input.longValue());
   }
 }
