+++ src/main/java/ar/com/kfgodel/function/intertype/longs/LongToFloatFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface LongToFloatFunction extends Function<Integer, Float> {
+public interface LongToFloatFunction extends Function<Long, Float> {
 
   float apply(long input);
 
   @Override
-  default Float apply(Integer input){
-    return apply(input.intValue());
+  default Float apply(Long input){
+    return apply(input.longValue());
   }
 }
