+++ src/main/java/ar/com/kfgodel/function/intertype/longs/LongToLongFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface LongToLongFunction extends Function<Integer, Long> {
+public interface LongToLongFunction extends Function<Long, Long> {
 
   long apply(long input);
 
   @Override
-  default Long apply(Integer input){
-    return apply(input.intValue());
+  default Long apply(Long input){
+    return apply(input.longValue());
   }
 }
