+++ src/main/java/ar/com/kfgodel/function/intertype/longs/LongToObjectFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface LongToObjectFunction<T> extends Function<Integer, T> {
+public interface LongToObjectFunction<T> extends Function<Long, T> {
 
   T apply(long input);
 
   @Override
-  default T apply(Integer input){
-    return apply(input.intValue());
+  default T apply(Long input){
+    return apply(input.longValue());
   }
 }
