+++ src/main/java/ar/com/kfgodel/function/intertype/longs/LongToShortFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface LongToShortFunction extends Function<Integer, Short> {
+public interface LongToShortFunction extends Function<Long, Short> {
 
   short apply(long input);
 
   @Override
-  default Short apply(Integer input){
-    return apply(input.intValue());
+  default Short apply(Long input){
+    return apply(input.longValue());
   }
 }
