+++ src/main/java/ar/com/kfgodel/function/intertype/shorts/ShortToByteFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface ShortToByteFunction extends Function<Integer, Byte> {
+public interface ShortToByteFunction extends Function<Short, Byte> {
 
   byte apply(short input);
 
   @Override
-  default Byte apply(Integer input){
-    return apply(input.intValue());
+  default Byte apply(Short input){
+    return apply(input.shortValue());
   }
 }
