+++ src/main/java/ar/com/kfgodel/function/intertype/shorts/ShortToDoubleFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface ShortToDoubleFunction extends Function<Integer, Double> {
+public interface ShortToDoubleFunction extends Function<Short, Double> {
 
   double apply(short input);
 
   @Override
-  default Double apply(Integer input){
-    return apply(input.intValue());
+  default Double apply(Short input){
+    return apply(input.shortValue());
   }
 }
