+++ src/main/java/ar/com/kfgodel/function/intertype/shorts/ShortToObjectFunction.java
@@ -5,12 +5,12 @@ import java.util.function.Function;
 /**
  * Date: 29/07/17 - 19:57
  */
-public interface ShortToObjectFunction<T> extends Function<Integer, T> {
+public interface ShortToObjectFunction<T> extends Function<Short, T> {
 
   T apply(short input);
 
   @Override
-  default T apply(Integer input){
-    return apply(input.intValue());
+  default T apply(Short input){
+    return apply(input.shortValue());
   }
 }
